+++ app/src/main/java/woowacourse/omok/Appplication.kt
@@ -1,0 +1,10 @@
+package woowacourse.omok
+
+import woowacourse.omok.controller.OmokController
+import woowacourse.omok.view.InputView
+import woowacourse.omok.view.OutputView
+
+fun main() {
+    val omokController = OmokController(InputView(), OutputView())
+    omokController.run()
+}
