+++ app/src/main/java/woowacourse/omok/model/Color.kt
@@ -1,4 +1,4 @@
-package omok.model
+package woowacourse.omok.model
 
 enum class Color {
     BLACK,
