+++ app/src/main/java/woowacourse/omok/model/Stone.kt
@@ -1,0 +1,8 @@
+package woowacourse.omok.model
+
+import woowacourse.omok.model.position.Position
+
+data class Stone(
+    val position: Position,
+    val color: Color,
+)
