+++ app/src/main/java/woowacourse/omok/model/position/Col.kt
@@ -1,4 +1,4 @@
-package omok.model.position
+package woowacourse.omok.model.position
 
 @JvmInline
 value class Col(val value: Int)
