+++ app/src/main/java/woowacourse/omok/model/position/Position.kt
@@ -1,4 +1,4 @@
-package omok.model.position
+package woowacourse.omok.model.position
 
 data class Position(
     val x: Col,
