+++ app/src/main/java/woowacourse/omok/model/position/Row.kt
@@ -1,4 +1,4 @@
-package omok.model.position
+package woowacourse.omok.model.position
 
 @JvmInline
 value class Row(val value: Int)
