--- app/src/main/kotlin/omok/model/Stone.kt
@@ -1,8 +1,0 @@
-package omok.model
-
-import omok.model.position.Position
-
-data class Stone(
-    val position: Position,
-    val color: Color,
-)
