+++ app/src/test/java/woowacourse/omok/model/RenjuRuleTest.kt
@@ -1,11 +1,11 @@
-package omok.model
+package woowacourse.omok.model
 
-import omok.model.position.Col
-import omok.model.position.Position
-import omok.model.position.Row
-import omok.model.rule.RenjuRule
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
+import woowacourse.omok.model.position.Col
+import woowacourse.omok.model.position.Position
+import woowacourse.omok.model.position.Row
+import woowacourse.omok.model.rule.RenjuRule
 
 class RenjuRuleTest {
     @Test
